hybrid-app/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
Werkzeug==3.0.3
